Modify Chat and Friend and User object model
