+++ Growing-Traveler/Growing-Traveler/Model/Profile/User.swift
@@ -1,0 +1,18 @@
+//
+//  User.swift
+//  Growing-Traveler
+//
+//  Created by Jenny Hung on 2022/4/19.
+//
+
+import Foundation
+
+struct User: Codable {
+    
+    var userID: String
+    
+    var userName: String
+    
+    var userEmail: String
+    
+}
